Add Windows compatibility.

+++ setup.py
@@ -10,10 +10,30 @@ from setuptools.command.develop import develop
 from distutils.command.clean import clean
 from distutils.core import Command
 from setuptools import setup, find_packages
+import platform
 
-_compile: Any = __import__('_compilers')
+def call_command(cmd):
+    try:
+        return subprocess.check_output(cmd, shell=True, stderr=subprocess.STDOUT).decode().strip()
+    except:
+        return None
 
-if _compile.is_windows:
+def get_version(compiler):
+    ver = call_command('gfortran -dumpversion')
+    print(ver)
+    if ver and '.' not in ver:
+        ver = call_command(compiler + ' -dumpfullversion')
+    return ver
+    
+import re
+
+def compare_version(version1, version2):
+    def normalize(v):
+        return [int(x) for x in re.sub(r'(\.0+)*$','', v).split(".")]
+        
+    return ((normalize(version1) > normalize(version2)) - (normalize(version1) < normalize(version2)))
+    
+if platform.system() == 'Window':
     DLLNAME = 'libcalc11.dll'
     DLLNAME_WRAP = 'libcalc11_wrapper.dll'
 else:
@@ -25,11 +45,27 @@ with open('README.md', "r") as fid:   #encoding='utf-8'
     
 
 def make_library():
-    #Check if windows work
+
+    #print(get_version('gcc'))
+    #print(get_version('gfortran'))
     
-    subprocess.call(('rm -rf calc11/lib/*.so'), shell=True)
-    os.chdir('./calc11/src')
-    subprocess.call(('rm -rf *.o','rm -rf *.so'), shell=True)
+    #print(platform.system())
+    #print(os.environ.copy())
+    
+    if get_version('gcc') is None:
+        raise IOError('gcc version 4.8.5 or later is required.')
+    
+    if compare_version(get_version('gcc'),'4.8.5') < 0:
+        raise IOError('gcc must be al least version 4.8.5.')
+    
+    if platform.system() == 'Windows':
+        subprocess.call(('rm -rf calc11\lib\*.so'), shell=True)
+        os.chdir('.\calc11\src')
+        subprocess.call(('del -rf *.o','del -rf *.so'), shell=True)
+    else:
+        subprocess.call(('rm -rf calc11/lib/*.so'), shell=True)
+        os.chdir('./calc11/src')
+        subprocess.call(('rm -rf *.o','rm -rf *.so'), shell=True)
     
     COMPILER_F = 'gfortran'
     FLAGS_F = '-ffree-form -ffree-line-length-none -fPIC -I. -g -O2 -c'
@@ -38,29 +74,35 @@ def make_library():
     cmd  = COMPILER_F + ' ' + FLAGS_F + ' ' + ' '.join([ s + '.f ' for s in SOURCES_F])
     os.system('pwd')
     if subprocess.call(cmd, shell=True) != 0:
-                    raise IOError('Compilation failed')
+        raise IOError('Compilation failed')
                     
     COMPILER_C = 'gcc'
     FLAGS_C = '-fPIC -c'
     SOURCES_C = ['almaout']
     cmd = COMPILER_C + ' ' + FLAGS_C + ' '  + ' '.join([s + '.c ' for s in SOURCES_C])
     if subprocess.call(cmd, shell=True) != 0:
-                    raise IOError('Compilation failed')
+        raise IOError('Compilation failed')
     
     SOURCES = SOURCES_F + SOURCES_C
     FLAGS = '-shared -o'
     cmd = COMPILER_F + ' ' + FLAGS + ' ' + DLLNAME + ' ' + ' '.join([ s + '.o ' for s in SOURCES])
     if subprocess.call(cmd, shell=True) != 0:
-                    raise IOError('Compilation failed')
+        raise IOError('Compilation failed')
                     
     cmd = COMPILER_F + ' -fPIC -shared -o calc11_wrapper.so calc11_wrapper.f90 -L. -lcalc11'
     if subprocess.call(cmd, shell=True) != 0:
-                    raise IOError('Compilation failed')
-                    
-    subprocess.call(('mv calc11_wrapper.so ../lib'), shell=True)
-    subprocess.call(('mv libcalc11.so ../lib'), shell=True)
+        raise IOError('Compilation failed')
+    
+    
+    if platform.system() == 'Windows':
+        subprocess.call(('move calc11_wrapper.so ../lib'), shell=True)
+        subprocess.call(('move libcalc11.so ../lib'), shell=True)
+    else:
+        subprocess.call(('mv calc11_wrapper.so ../lib'), shell=True)
+        subprocess.call(('mv libcalc11.so ../lib'), shell=True)
     
     os.chdir('../../')
+    
     
 #    cmd = 'make -C ./calc11/src/ clean'
 #    if subprocess.call(cmd, shell=True, env=_compile.compiler_environ) != 0:
